park-runs/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
django-cors-headers==4.3.1
